+++ kit/ChildSession.hpp
@@ -23,6 +23,7 @@
 #include "Kit.hpp"
 #include "Session.hpp"
 #include "Watermark.hpp"
+#include "StateRecorder.hpp"
 
 class ChildSession;
 
@@ -32,181 +33,7 @@ enum class LokEventTargetEnum
     Window
 };
 
-// An abstract interface.
-class DocumentManagerInterface
-{
-public:
-    virtual ~DocumentManagerInterface()  {}
-
-    /// Request loading a document, or a new view, if one exists.
-    virtual bool onLoad(const std::string& sessionId,
-                        const std::string& uriAnonym,
-                        const std::string& renderOpts) = 0;
-
-    /// Unload a client session, which unloads the document
-    /// if it is the last and only.
-    virtual void onUnload(const ChildSession& session) = 0;
-
-    /// Access to the Kit instance.
-    virtual std::shared_ptr<lok::Office> getLOKit() = 0;
-
-    /// Access to the document instance.
-    virtual std::shared_ptr<lok::Document> getLOKitDocument() = 0;
-
-    /// Send msg to all active sessions.
-    virtual bool notifyAll(const std::string& msg) = 0;
-
-    /// Send updated view info to all active sessions.
-    virtual void notifyViewInfo() = 0;
-    virtual void updateEditorSpeeds(int id, int speed) = 0;
-
-    virtual int getEditorId() const = 0;
-
-    /// Get a view ID <-> UserInfo map.
-    virtual std::map<int, UserInfo> getViewInfo() = 0;
-
-    virtual std::string getObfuscatedFileId() = 0;
-
-    virtual std::shared_ptr<TileQueue>& getTileQueue() = 0;
-
-    virtual bool sendFrame(const char* buffer, int length, WSOpCode opCode = WSOpCode::Text) = 0;
-
-    virtual void alertAllUsers(const std::string& cmd, const std::string& kind) = 0;
-
-    virtual unsigned getMobileAppDocId() const = 0;
-
-    /// See if we should clear out our memory
-    virtual void trimIfInactive() = 0;
-
-    virtual bool isDocPasswordProtected() const = 0;
-
-    virtual bool haveDocPassword() const = 0;
-
-    virtual std::string getDocPassword() const = 0;
-
-    virtual DocumentPasswordType getDocPasswordType() const = 0;
-
-    virtual void updateActivityHeader() const = 0;
-};
-
-struct RecordedEvent
-{
-private:
-    int _type = 0;
-    std::string _payload;
-
-public:
-    RecordedEvent()
-    {
-    }
-
-    RecordedEvent(int type, const std::string& payload)
-        : _type(type),
-        _payload(payload)
-    {
-    }
-
-    void setType(int type)
-    {
-        _type = type;
-    }
-
-    int getType() const
-    {
-        return _type;
-    }
-
-    void setPayload(const std::string& payload)
-    {
-        _payload = payload;
-    }
-
-    const std::string& getPayload() const
-    {
-        return _payload;
-    }
-};
-
-/// When the session is inactive, we need to record its state for a replay.
-class StateRecorder
-{
-private:
-    bool _invalidate;
-    std::unordered_map<std::string, std::string> _recordedStates;
-    std::unordered_map<int, std::unordered_map<int, RecordedEvent>> _recordedViewEvents;
-    std::unordered_map<int, RecordedEvent> _recordedEvents;
-    std::vector<RecordedEvent> _recordedEventsVector;
-
-public:
-    StateRecorder() : _invalidate(false) {}
-
-    // TODO Remember the maximal area we need to invalidate - grow it step by step.
-    void recordInvalidate()
-    {
-        _invalidate = true;
-    }
-
-    bool isInvalidate() const
-    {
-        return _invalidate;
-    }
-
-    const std::unordered_map<std::string, std::string>& getRecordedStates() const
-    {
-        return _recordedStates;
-    }
-
-    const std::unordered_map<int, std::unordered_map<int, RecordedEvent>>& getRecordedViewEvents() const
-    {
-        return _recordedViewEvents;
-    }
-
-    const std::unordered_map<int, RecordedEvent>& getRecordedEvents() const
-    {
-        return _recordedEvents;
-    }
-
-    const std::vector<RecordedEvent>& getRecordedEventsVector() const
-    {
-        return _recordedEventsVector;
-    }
-
-    void recordEvent(const int type, const std::string& payload)
-    {
-        _recordedEvents[type] = RecordedEvent(type, payload);
-    }
-
-    void recordViewEvent(const int viewId, const int type, const std::string& payload)
-    {
-        _recordedViewEvents[viewId][type] = {type, payload};
-    }
-
-    void recordState(const std::string& name, const std::string& value)
-    {
-        _recordedStates[name] = value;
-    }
-
-    /// In the case we need to remember all the events that come, not just
-    /// the final state.
-    void recordEventSequence(const int type, const std::string& payload)
-    {
-        _recordedEventsVector.emplace_back(type, payload);
-    }
-
-    void dumpState(std::ostream&)
-    {
-        // TODO: the rest ...
-    }
-
-    void clear()
-    {
-        _invalidate = false;
-        _recordedEvents.clear();
-        _recordedViewEvents.clear();
-        _recordedStates.clear();
-        _recordedEventsVector.clear();
-    }
-};
+class DocumentManagerInterface;
 
 /// Represents a session to the WSD process, in a Kit process. Note that this is not a singleton.
 class ChildSession final : public Session
